requirements.txt
fastapi
uvicorn
pydantic
python-multipart
firebase-admin
google-generativeai
python-dotenv
python-pptx
Pillow
trimesh
numpy
requests
